+++ src/main/java/com/example/Abacus/Repo/EnquiryRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.Abacus.Model.Enquiry;
 
 @Repository
-public interface EnquiryRepo extends JpaRepository<Enquiry, Long> {
+public interface EnquiryRepo extends JpaRepository<Enquiry, Integer> {
     
 }
